Create censoring bronze
First approach was to just to do find function and loop until there is no more. 
Problem was that the test cases were too large and encountered an time error. solution was to just to iterate it through and everytime the end substring   = the censored word, just delete the end and continue. This way, we can solve it by just iterating once instead of iteraing over and then having a nested function